+++ features/step_definitions/error_replies.rb
@@ -73,10 +73,6 @@ Then /^the server returns a format not implemented error$/ do
   step 'the server returns a "504 Format not implemented" error'
 end
 
-Then /^the server returns an invalid format error$/ do
-  step 'the server returns a "504 Invalid format code" error'
-end
-
 Then /^the server returns a command unrecognized error$/ do
   step 'the server returns a "500 Syntax error, command unrecognized" error'
 end