+++ Changelog.md
@@ -12,7 +12,8 @@ Enhancements
 * Added --auth switch to the example to select the authentication
   level.
 * Support APPE
-* Support TYPE A T (ASCII/Telnet)
+* Support TYPE "A T" (ASCII/Telnet)
+* Support TYPE "LOCAL 8"
 
 API changes